[#176236216] - Split subscribe off into submodule
